# Performance: Speed up NotWritten outdatedness rule