### Detailed description

This changes the `NotWritten` to pre-load the list of files in the output directory, so that the filesystem only has to be accessed once, rather than for every file.

This leads to a nice speedup. For one particular site:

```
                                          x │ count     min     .50     .90     .95     max     tot
────────────────────────────────────────────┼──────────────────────────────────────────────────────
 Nanoc::Core::OutdatednessRules::NotWritten │  3704   0.00s   0.00s   0.01s   0.01s   0.05s   4.77s
```

```
                                          x │ count     min     .50     .90     .95     max     tot
────────────────────────────────────────────┼──────────────────────────────────────────────────────
 Nanoc::Core::OutdatednessRules::NotWritten │  3704   0.00s   0.00s   0.00s   0.00s   0.05s   0.07s
```

Before: 4.77s. After: 0.07s. That’s a 60x–70x speedup of the `NotWritten` rule.

### To do

(Include the to-do list for this PR to be finished here.)

* [x] Tests
* [x] Check in output directory, not in `Dir.pwd`

Future work could entail excluding directories such as `.git/`, but that’d likely require switching from `Dir.glob` to `Find`, which might not yield much of a speedup. In any case, the `NotWritten` rule is no longer a bottleneck now.

### Related issues

n/a