@opatry 2022-10-01
Sounds nice!

I'm quite interested by this kind of improvements given my personal photoblog being quite heavy in terms of files in output.

Can you elaborate a bit on use cases where your change can have impact?

Eg on my side:
- 3554 generated file(s)
  - 550 html
  - 1487 jpg
  - 1487 webp
  - other