@denisdefreyne 2022-10-01
@opatry Interestingly, now that I’m re-running the performance tests for this change, I see a much smaller benefit.

I had some trouble with my MacBook Air M2 being throttled, presumably because it ran quite hot.

I don’t get results with `File.file?` that are even *close* to 4.77s anymore. Old `File.file?` approach: ~0.04s. New `Dir.glob` approach: ~0.02s.

I’ll have to do some more thorough testing, but this might have not had nearly as much of an impact as I had hoped.